style(entity,dto,service,...): 实体等pojo主键变为Long类型
主要对article实体等pojo主键变为Long类型，并对其涉及到的代码作一系列适配
article部分代码对service实现返回map做一系列修正

+++ src/main/java/com/rymcu/forest/mapper/FollowMapper.java
@@ -18,19 +18,19 @@ public interface FollowMapper extends Mapper<Follow> {
      * @param followingType
      * @return
      */
-    Boolean isFollow(@Param("followingId") Integer followingId, @Param("followerId") Integer followerId, @Param("followingType") String followingType);
+    Boolean isFollow(@Param("followingId") Integer followingId, @Param("followerId") Long followerId, @Param("followingType") String followingType);
 
     /**
      * 查询用户粉丝
      * @param idUser
      * @return
      */
-    List<UserDTO> selectUserFollowersByUser(@Param("idUser") Integer idUser);
+    List<UserDTO> selectUserFollowersByUser(@Param("idUser") Long idUser);
 
     /**
      * 查询用户关注用户
      * @param idUser
      * @return
      */
-    List<UserDTO> selectUserFollowingsByUser(@Param("idUser") Integer idUser);
+    List<UserDTO> selectUserFollowingsByUser(@Param("idUser") Long idUser);
 }
